requirements.txt
discord.py>=2.7.1
Pillow>=12.2.0
